Refactoring of tests
