+++ src/test/java/com/remondis/remap/basic/AWithOneMoreSourceField.java
@@ -1,4 +1,4 @@
-package com.remondis.remap;
+package com.remondis.remap.basic;
 
 public class AWithOneMoreSourceField {
