+++ src/test/java/com/remondis/remap/collections/A.java
@@ -5,7 +5,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import com.remondis.remap.B;
+import com.remondis.remap.basic.B;
 
 public class A {
