+++ src/test/java/com/remondis/remap/inheritance/Child.java
@@ -1,6 +1,6 @@
 package com.remondis.remap.inheritance;
 
-import com.remondis.remap.B;
+import com.remondis.remap.basic.B;
 
 public class Child extends Parent {
