+++ src/test/java/com/remondis/remap/inheritance/MapperTest.java
@@ -6,10 +6,10 @@ import static org.junit.Assert.assertNull;
 
 import org.junit.Test;
 
-import com.remondis.remap.B;
-import com.remondis.remap.BResource;
 import com.remondis.remap.Mapper;
 import com.remondis.remap.Mapping;
+import com.remondis.remap.basic.B;
+import com.remondis.remap.basic.BResource;
 
 public class MapperTest {
